+++ packages/terra-date-picker/src/react-datepicker/week.jsx
@@ -43,6 +43,11 @@ export default class Week extends React.Component {
      */
     includeDates: PropTypes.array,
     /**
+     * Timezone value to indicate in which timezone the date-time component is rendered.
+     * The value provided should be a valid [timezone](https://en.wikipedia.org/wiki/List_of_tz_database_time_zones) string, else will default to browser/local timezone.
+     */
+    initialTimeZone: PropTypes.string,
+    /**
      * Prop to show inline version of date picker component.
      */
     inline: PropTypes.bool,
@@ -103,10 +108,6 @@ export default class Week extends React.Component {
      * Minimum date for a given range.
      */
     startDate: PropTypes.object,
-    /**
-     * Difference between utc and local time.
-     */
-    utcOffset: PropTypes.number,
     /**
      * Whether or not calendar is opened via keyboard
      */
@@ -169,6 +170,7 @@ export default class Week extends React.Component {
           maxDate={this.props.maxDate}
           excludeDates={this.props.excludeDates}
           includeDates={this.props.includeDates}
+          initialTimeZone={this.props.initialTimeZone}
           inline={this.props.inline}
           highlightDates={this.props.highlightDates}
           selectingDate={this.props.selectingDate}
@@ -179,8 +181,7 @@ export default class Week extends React.Component {
           selectsEnd={this.props.selectsEnd}
           startDate={this.props.startDate}
           endDate={this.props.endDate}
-          dayClassName={this.props.dayClassName}
-          utcOffset={this.props.utcOffset}/>
+          dayClassName={this.props.dayClassName}/>
       )
     }))
   }